Move exceptions stored in the /type directory to /exceptions
