+++ src/test/java/edu/hawaii/its/api/type/GroupingServiceResultExceptionTest.java
@@ -3,6 +3,8 @@ package edu.hawaii.its.api.type;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
 
+import edu.hawaii.its.api.exception.GroupingsServiceResultException;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
