+++ src/test/java/edu/hawaii/its/api/type/GroupingsHTTPExceptionTest.java
@@ -2,6 +2,8 @@ package edu.hawaii.its.api.type;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import edu.hawaii.its.api.exception.GroupingsHTTPException;
+
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
